movement of validate_ssh_pub_key to utils
