+++ robottelo/host_helpers/satellite_mixins.py
@@ -95,9 +95,7 @@ class ContentInfo:
             reached or calculation was not successful).
         """
         filename = url.split('/')[-1]
-        result = self.execute(
-            f'wget -qO - {url} | tee {filename} | md5sum | awk \'{{print $1}}\''
-        )
+        result = self.execute(f'wget -qO - {url} | tee {filename} | md5sum | awk \'{{print $1}}\'')
         if result.status != 0:
             raise AssertionError(f'Failed to calculate md5 checksum of {filename}')
         return result.stdout
